test: refactor test arrangment to base class
